Move back some tests from norace_2_test.go to norace_1_test.go
